CoxExample {glmnet}	R Documentation

Synthetic dataset with right-censored survival response

Description

Randomly generated data for Cox regression example.

Usage

data(CoxExample)

Format

List containing the following elements:

x

1,000 by 30 matrix of numeric values.

y

1,000 by 2 matrix with column names "time" and "status". The first column consists of positive numbers representing time to event, while the second column represents the status indicator (0=right-censored, 1=observed).

[Package glmnet version 4.1-10 Index]