SparseExample {glmnet}	R Documentation

Synthetic dataset with sparse design matrix

Description

Randomly generated data for Gaussian regression example with the design matrix x being in sparse matrix format.

Usage

data(SparseExample)

Format

List containing the following elements:

x

100 by 20 matrix of numeric values. x is in sparse matrix format, having class "dgCMatrix".

y

Numeric vector of length 100.

[Package glmnet version 4.1-10 Index]