cox_obj_function {glmnet}	R Documentation

Elastic net objective function value for Cox regression model

Description

Returns the elastic net objective function value for Cox regression model.

Usage

cox_obj_function(y, pred, weights, lambda, alpha, coefficients, vp)

Arguments

y
Survival response variable, must be a Surv or stratifySurv object.

pred
Model's predictions for y.

weights
Observation weights.

lambda
A single value for the lambda hyperparameter.

alpha
The elasticnet mixing parameter, with 0 ≤ α ≤ 1.

coefficients
The model's coefficients.

vp
Penalty factors for each of the coefficients.

[Package glmnet version 4.1-10 Index]