dev_function {glmnet}	R Documentation

Elastic net deviance value

Description

Returns the elastic net deviance value.

Usage

dev_function(y, mu, weights, family)

Arguments

y
Quantitative response variable.

mu
Model's predictions for y.

weights
Observation weights.

family
A description of the error distribution and link function to be used in the model. This is the result of a call to a family function.

[Package glmnet version 4.1-10 Index]