fid {glmnet}	R Documentation

Helper function for Cox deviance and gradient

Description

Helps to find ties in death times of data.

Usage

fid(x, index)

Arguments

x
Sorted vector of death times.

index
Vector of indices for the death times.

Value

A list with two arguments.

index_first
A vector of indices for the first observation at each death time as they appear in the sorted list.

index_ties
If there are no ties at all, this is NULL. If not, this is a list with length equal to the number of unique times with ties. For each time with ties, index_ties gives the indices of the observations with a death at that time.

Examples

# Example with no ties
glmnet:::fid(c(1, 4, 5, 6), 1:5)

# Example with ties
glmnet:::fid(c(1, 1, 1, 2, 3, 3, 4, 4, 4), 1:9)

[Package glmnet version 4.1-10 Index]