glmnet-internal {glmnet}	R Documentation

Internal glmnet functions

Description

Internal glmnet functions

Usage

fix.lam(lam)
lambda.interp(lambda,s)
nonzeroCoef(beta, bystep=FALSE)
glmnet_softmax(x)

Details

These are not intended for use by users. lambda.interp does linear interpolation of the lambdas to obtain a prediction at a new point s. glmnet_softmax does the classification for multinomial models. nonzeroCoef determines in an efficient manner which variables are nonzero in each fit. jerr prints out error messages from the fortran. plotCoef is called by the plot method for glmnet objects.

Author(s)

Trevor Hastie

[Package glmnet version 1.9-5 Index]