internal glmnet parameters

View and/or change the factory default parameters in glmnet

glmnet.control(fdev = 1e-05, devmax = 0.999, eps = 1e-06,
  big = 9.9e+35, mnlam = 5, pmin = 1e-09, exmx = 250,
  prec = 1e-10, mxit = 100, itrace = 0, factory = FALSE)

Arguments

fdev
minimum fractional change in deviance for stopping path; factory default = 1.0e-5

devmax
maximum fraction of explained deviance for stopping path; factory default = 0.999

eps
minimum value of lambda.min.ratio (see glmnet); factory default= 1.0e-6

big
large floating point number; factory default = 9.9e35. Inf in definition of upper.limit is set to big

mnlam
minimum number of path points (lambda values) allowed; factory default = 5

pmin
minimum probability for any class. factory default = 1.0e-9. Note that this implies a pmax of 1-pmin.

exmx
maximum allowed exponent. factory default = 250.0

prec
convergence threshold for multi response bounds adjustment solution. factory default = 1.0e-10

mxit
maximum iterations for multiresponse bounds adjustment solution. factory default = 100

itrace
If 1 then progress bar is displayed when running glmnet and cv.glmnet. factory default = 0

factory
If TRUE, reset all the parameters to the factory default; default is FALSE

Value

A list with named elements as in the argument list

Details

If called with no arguments, glmnet.control() returns a list with the current settings of these parameters. Any arguments included in the call sets those parameters to the new values, and then silently returns. The values set are persistent for the duration of the R session.

See also

Examples

glmnet.control(fdev = 0)  #continue along path even though not much changes
glmnet.control()  # view current settings

#> $fdev
#> [1] 0
#> 
#> $eps
#> [1] 1e-06
#> 
#> $big
#> [1] 9.9e+35
#> 
#> $mnlam
#> [1] 5
#> 
#> $devmax
#> [1] 0.999
#> 
#> $pmin
#> [1] 1e-09
#> 
#> $exmx
#> [1] 250
#> 
#> $itrace
#> [1] 0
#> 
#> $prec
#> [1] 1e-10
#> 
#> $mxit
#> [1] 100
#> 

glmnet.control(factory = TRUE)  # reset all the parameters to their default