mycoxph {glmnet}	R Documentation

Helper function to fit coxph model for survfit.coxnet

Description

This function constructs the coxph call needed to run the "hack" of coxph with 0 iterations. It's a separate function as we have to deal with function options like strata, offset and observation weights.

Usage

mycoxph(object, s, ...)

Arguments

object
A class coxnet object.

s
The value of the penalty parameter lambda at which the survival curve is required.

...
The same ... that was passed to survfit.coxnet.

[Package glmnet version 4.1-10 Index]