mycoxpred {glmnet}	R Documentation

Helper function to amend ... for new data in survfit.coxnet

Description

This function amends the function arguments passed to survfit.coxnet via ... if new data was passed to survfit.coxnet. It's a separate function as we have to deal with function options like newstrata and newoffset.

Usage

mycoxpred(object, s, ...)

Arguments

object
A class coxnet object.

s
The response for the fitted model.

...
The same ... that was passed to survfit.coxnet.

[Package glmnet version 4.1-10 Index]