print a glmnet object

Print a summary of the glmnet path at each step along the path.

# S3 method for glmnet
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

x
fitted glmnet object

digits
significant digits in printout

...
additional print arguments

Value

The matrix above is silently returned

Details

The call that produced the object x is printed, followed by a three-column matrix with columns Df, %Dev and Lambda. The Df column is the number of nonzero coefficients (Df is a reasonable name only for lasso fits). %Dev is the percent deviance explained (relative to the null deviance). In the case of a 'relaxed' fit, an additional column is inserted, %Dev R which gives the percent deviance explained by the relaxed model. For a "bigGlm" model, a simpler summary is printed.

References

Friedman, J., Hastie, T. and Tibshirani, R. (2008). Regularization Paths for Generalized Linear Models via Coordinate Descent

See also

Examples

x = matrix(rnorm(100 * 20), 100, 20)
y = rnorm(100)
fit1 = glmnet(x, y)
print(fit1)

#> 
#> Call:  glmnet(x = x, y = y) 
#> 
#>    Df    %Dev   Lambda
#> 1   0 0.00000 0.209400
#> 2   2 0.01238 0.190800
#> 3   3 0.02575 0.173800
#> 4   3 0.03886 0.158400
#> 5   4 0.05062 0.144300
#> 6   6 0.06705 0.131500
#> 7   8 0.08540 0.119800
#> 8   8 0.10250 0.109200
#> 9   8 0.11670 0.099460
#> 10  8 0.12850 0.090630
#> 11  8 0.13830 0.082580
#> 12  8 0.14640 0.075240
#> 13  8 0.15320 0.068560
#> 14  9 0.15970 0.062470
#> 15 10 0.16530 0.056920
#> 16 13 0.17080 0.051860
#> 17 13 0.17620 0.047250
#> 18 13 0.18080 0.043060
#> 19 13 0.18460 0.039230
#> 20 14 0.18770 0.035750
#> 21 16 0.19090 0.032570
#> 22 16 0.19360 0.029680
#> 23 16 0.19580 0.027040
#> 24 16 0.19770 0.024640
#> 25 16 0.19920 0.022450
#> 26 17 0.20070 0.020460
#> 27 18 0.20200 0.018640
#> 28 18 0.20320 0.016980
#> 29 18 0.20410 0.015470
#> 30 18 0.20490 0.014100
#> 31 19 0.20560 0.012850
#> 32 19 0.20620 0.011710
#> 33 19 0.20670 0.010670
#> 34 19 0.20710 0.009718
#> 35 19 0.20740 0.008854
#> 36 19 0.20770 0.008068
#> 37 19 0.20790 0.007351
#> 38 19 0.20810 0.006698
#> 39 19 0.20830 0.006103
#> 40 19 0.20840 0.005561
#> 41 19 0.20860 0.005067
#> 42 19 0.20860 0.004617
#> 43 19 0.20870 0.004207
#> 44 19 0.20880 0.003833
#> 45 19 0.20880 0.003492
#> 46 19 0.20890 0.003182
#> 47 19 0.20890 0.002899
#> 48 19 0.20890 0.002642
#> 49 20 0.20900 0.002407
#> 50 20 0.20900 0.002193
#> 51 20 0.20900 0.001998
#> 52 20 0.20900 0.001821
#> 53 20 0.20900 0.001659
#> 54 20 0.20910 0.001512
#> 55 20 0.20910 0.001377
#> 56 20 0.20910 0.001255
#> 57 20 0.20910 0.001144
#> 58 20 0.20910 0.001042
#> 59 20 0.20910 0.000949
#> 60 20 0.20910 0.000865
#> 61 20 0.20910 0.000788
#> 62 20 0.20910 0.000718
#> 63 20 0.20910 0.000654
#> 64 20 0.20910 0.000596