rmult {glmnet}	R Documentation

Generate multinomial samples from a probability matrix

Description

Generate multinomial samples

Usage

rmult(p)

Arguments

p
matrix of probabilities, with number of columns the number of classes

Details

Simple function that calls the rmultinom function. It generates a class label for each row of its input matrix of class probabilities.

Value

a vector of class memberships

Author(s)

Trevor Hastie
Maintainer: Trevor Hastie hastie@stanford.edu

[Package glmnet version 4.1-10 Index]