compute C index for a Cox model

Computes Harrel's C index for predictions from a "coxnet" object.

Cindex(pred, y, weights = rep(1, nrow(y)))

Arguments

pred
Predictions from a "coxnet" object

y
a survival response object - a matrix with two columns "time" and "status"; see documentation for "glmnet"

weights
optional observation weights

Details

Computes the concordance index, taking into account censoring.

References

Harrel Jr, F. E. and Lee, K. L. and Mark, D. B. (1996) Tutorial in biostatistics: multivariable prognostic models: issues in developing models, evaluating assumptions and adequacy, and measuring and reducing error, Statistics in Medicine, 15, pages 361--387.

See also

Examples

set.seed(10101)
N = 1000
p = 30
nzc = p/3
x = matrix(rnorm(N * p), N, p)
beta = rnorm(nzc)
fx = x[, seq(nzc)] %*% beta/3
hx = exp(fx)
ty = rexp(N, hx)
tcens = rbinom(n = N, prob = 0.3, size = 1)  # censoring indicator
y = cbind(time = ty, status = 1 - tcens)  # y=Surv(ty,1-tcens) with library(survival)
fit = glmnet(x, y, family = "cox")
pred = predict(fit, newx = x)
Cindex(pred, y)

#> [1] 0.5

cv.glmnet(x, y, family = "cox", type.measure = "C")

#> 
#> Call:  cv.glmnet(x = x, y = y, type.measure = "C", family = "cox") 
#> 
#> Measure: C-index 
#> 
#>      Lambda Measure      SE Nonzero
#> min 0.01920  0.7269 0.01170      14
#> 1se 0.08509  0.7154 0.01095       8