beta_CVX {glmnet}	R Documentation

Simulated data for the glmnet vignette

Description

Simple simulated data, used to demonstrate the features of glmnet

Usage

data(BinomialExample)
data(CVXResults)
data(CoxExample)
data(MultiGaussianExample)
data(MultinomialExample)
data(PoissonExample)
data(QuickStartExample)
data(SparseExample)

Format

Data objects used to demonstrate features in the glmnet vignette

Details

Theses datasets are artificial, and ere used to test out some of the features of glmnet.

Examples

data(QuickStartExample)
glmnet(x,y)

[Package glmnet version 2.0-16 Index]