Simulated data for the glmnet vignette

Simple simulated data, used to demonstrate the features of glmnet

Format

Data objects used to demonstrate features in the glmnet vignette

Details

These datasets are artificial, and are used to test out some of the features of glmnet.

Examples

data(QuickStartExample)
glmnet(x, y)

#> 
#> Call:  glmnet(x = x, y = y) 
#> 
#>    Df    %Dev  Lambda
#> 1   0 0.00000 1.63100
#> 2   2 0.05528 1.48600
#> 3   2 0.14590 1.35400
#> 4   2 0.22110 1.23400
#> 5   2 0.28360 1.12400
#> 6   2 0.33540 1.02400
#> 7   4 0.39040 0.93320
#> 8   5 0.45600 0.85030
#> 9   5 0.51540 0.77470
#> 10  6 0.57350 0.70590
#> 11  6 0.62550 0.64320
#> 12  6 0.66870 0.58610
#> 13  6 0.70460 0.53400
#> 14  6 0.73440 0.48660
#> 15  7 0.76210 0.44330
#> 16  7 0.78570 0.40400
#> 17  7 0.80530 0.36810
#> 18  7 0.82150 0.33540
#> 19  7 0.83500 0.30560
#> 20  7 0.84620 0.27840
#> 21  7 0.85550 0.25370
#> 22  7 0.86330 0.23120
#> 23  8 0.87060 0.21060
#> 24  8 0.87690 0.19190
#> 25  8 0.88210 0.17490
#> 26  8 0.88650 0.15930
#> 27  8 0.89010 0.14520
#> 28  8 0.89310 0.13230
#> 29  8 0.89560 0.12050
#> 30  8 0.89760 0.10980
#> 31  9 0.89940 0.10010
#> 32  9 0.90100 0.09117
#> 33  9 0.90230 0.08307
#> 34  9 0.90340 0.07569
#> 35 10 0.90430 0.06897
#> 36 11 0.90530 0.06284
#> 37 11 0.90620 0.05726
#> 38 12 0.90700 0.05217
#> 39 15 0.90780 0.04754
#> 40 16 0.90860 0.04331
#> 41 16 0.90930 0.03947
#> 42 16 0.90980 0.03596
#> 43 17 0.91030 0.03277
#> 44 17 0.91070 0.02985
#> 45 18 0.91110 0.02720
#> 46 18 0.91140 0.02479
#> 47 19 0.91170 0.02258
#> 48 19 0.91200 0.02058
#> 49 19 0.91220 0.01875
#> 50 19 0.91240 0.01708
#> 51 19 0.91250 0.01557
#> 52 19 0.91260 0.01418
#> 53 19 0.91270 0.01292
#> 54 19 0.91280 0.01178
#> 55 19 0.91290 0.01073
#> 56 19 0.91290 0.00978
#> 57 19 0.91300 0.00891
#> 58 19 0.91300 0.00812
#> 59 19 0.91310 0.00740
#> 60 19 0.91310 0.00674
#> 61 19 0.91310 0.00614
#> 62 20 0.91310 0.00559
#> 63 20 0.91310 0.00510
#> 64 20 0.91310 0.00464
#> 65 20 0.91320 0.00423
#> 66 20 0.91320 0.00386
#> 67 20 0.91320 0.00351