glmnet-internal {glmnet}	R Documentation

Internal glmnet functions

Description

Internal glmnet functions

Usage

fix.lam(lam)
lambda.interp(lambda,s)
nonzeroCoef(beta, bystep=FALSE)
glmnet_softmax(x)

Details

These are not intended for use by users. lambda.interp does linear interpolation of the lambdas to obtain a prediction at a new point s. glmnet_softmax does the classification for multinomial models. nonzeroCoef determines in an efficient manner which variables are nonzero in each fit. jerr prints out error messages from the fortran. plotCoef is called by the plot method for glmnet objects. check_dots is used in coef and predict with argument exact=TRUE, to make sure user supplies original data used to fit the "glmnet" object.

Author(s)

Trevor Hastie

[Package glmnet version 2.0-16 Index]