glmnet.control {glmnet}	R Documentation

internal glmnet parameters

Description

View and/or change the factory default parameters in glmnet

Usage

glmnet.control(fdev=1.0e-5, devmax=0.999, eps=1.0e-6, big=9.9e35, mnlam=5, pmin=1.0e-9,
     exmx=250.0,prec=1e-10,mxit=100,factory=FALSE)

Arguments

fdev
minimum fractional change in deviance for stopping path; factory default = 1.0e-5

devmax
maximum fraction of explained deviance for stopping path; factory default = 0.999

eps
minimum value of lambda.min.ratio (see glmnet); factory default= 1.0e-6

big
large floating point number; factory default = 9.9e35. Inf in definition of upper.limit is set to big

mnlam
minimum number of path points (lambda values) allowed; factory default = 5

pmin
minimum probability for any class. factory default = 1.0e-9. Note that this implies a pmax of 1-pmin.

exmx
maximum allowed exponent. factory default = 250.0

prec
convergence threshold for multi response bounds adjustment solution. factory default = 1.0e-10

mxit
maximum iterations for multiresponse bounds adjustment solution. factory default = 100

factory
If TRUE, reset all the parameters to the factory default; default is FALSE

Details

If called with no arguments, glmnet.control() returns a list with the current settings of these parameters. Any arguments included in the call sets those parameters to the new values, and then silently returns. The values set are persistent for the duration of the R session.

Value

A list with named elements as in the argument list

Author(s)

Jerome Friedman, Trevor Hastie
Maintainer: Trevor Hastie hastie@stanford.edu

See Also

glmnet

Examples

glmnet.control(fdev=0)#continue along path even though not much changes
glmnet.control() # view current settings
glmnet.control(factory=TRUE) # reset all the parameters to their default

[Package glmnet version 2.0-16 Index]