predict.cv.glmnet {glmnet}	R Documentation

make predictions from a "cv.glmnet" object.

Description

This function makes predictions from a cross-validated glmnet model, using the stored "glmnet.fit" object, and the optimal value chosen for lambda.

Usage

## S3 method for class 'cv.glmnet'
predict(object, newx, s=c("lambda.1se","lambda.min"),...)
## S3 method for class 'cv.glmnet'
coef(object,s=c("lambda.1se","lambda.min"),...)

Arguments

object
Fitted "cv.glmnet" object.

newx
Matrix of new values for x at which predictions are to be made. Must be a matrix; can be sparse as in Matrix package. See documentation for predict.glmnet.

s
Value(s) of the penalty parameter lambda at which predictions are required. Default is the value s="lambda.1se" stored on the CV object. Alternatively s="lambda.min" can be used. If s is numeric, it is taken as the value(s) of lambda to be used.

...
Not used. Other arguments to predict.

Details

This function makes it easier to use the results of cross-validation to make a prediction.

Value

The object returned depends the ... argument which is passed on to the predict method for glmnet objects.

Author(s)

Jerome Friedman, Trevor Hastie and Rob Tibshirani
Maintainer: Trevor Hastie <hastie@stanford.edu>

References

Friedman, J., Hastie, T. and Tibshirani, R. (2008) Regularization Paths for Generalized Linear Models via Coordinate Descent, Journal of Statistical Software, Vol. 33, Issue 1, Feb 2010
http://www.jstatsoft.org/v33/i01/

See Also

glmnet, and print, and coef methods, and cv.glmnet.

Examples

x=matrix(rnorm(100*20),100,20)
y=rnorm(100)
cv.fit=cv.glmnet(x,y)
predict(cv.fit,newx=x[1:5,])
coef(cv.fit)
coef(cv.fit,s="lambda.min")
predict(cv.fit,newx=x[1:5,],s=c(0.001,0.002))

[Package glmnet version 2.0-16 Index]